split Scatter
